requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
pillow==10.4.0
diffusers==0.30.0
transformers==4.42.0
accelerate==0.33.0
safetensors==0.4.3
